Reorganize plugin code, start defining API
There are mainly 5 different logic blocks:
- The plugin entrypoint, which creates necessary folders and files.
- The model, which contains a couple of helper classes to pass data
  around.
- The screenshot comparers, both single and multithreaded versions.
- The "base screenshot test", which should be changed and renamed into
  some kind of runner or framework class. It should be the only exposed
  class for people using the library, so it can be used in GUT tests or
  regular GDScript code.
- The config class, which right now only contains the folders' paths,
  but it should contain other config options (eg. number of shards,
  whether it's a regenerate base screenshots run or not, etc.).

The API still needs to be defined. Right now the "base screenshot test"
is how it has to be used, because the API is not defined. Ideally people
using this library should be able to just `take_screenshot("name")` and,
just by checking in the config, it should decide whether it should
compare the base screenshot with the new one, or regenerate the base.