Implement test runner and test report
Implements a simple test runner that can be called from the terminal and
a test report generator that generates a "junit.xml" style report.

It also includes new examples to show how to use the Base Screeshot Test
or the API calls directly (in case someone wants to use GUT as a test
runner).